Code clean and refactoring.
