+++ src/main/java/org/jboss/logging/processor/apt/MessageMethodBuilder.java
@@ -47,11 +47,11 @@ import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
 
 import org.jboss.logging.processor.apt.Annotations.FormatType;
-import org.jboss.logging.processor.intf.model.MessageMethod;
-import org.jboss.logging.processor.intf.model.Parameter;
-import org.jboss.logging.processor.intf.model.Parameter.ParameterType;
-import org.jboss.logging.processor.intf.model.ReturnType;
-import org.jboss.logging.processor.intf.model.ThrowableType;
+import org.jboss.logging.processor.model.MessageMethod;
+import org.jboss.logging.processor.model.Parameter;
+import org.jboss.logging.processor.model.Parameter.ParameterType;
+import org.jboss.logging.processor.model.ReturnType;
+import org.jboss.logging.processor.model.ThrowableType;
 import org.jboss.logging.processor.util.Comparison;
 
 /**
