+++ src/main/java/org/jboss/logging/processor/apt/ThrowableTypeFactory.java
@@ -22,7 +22,7 @@
 
 package org.jboss.logging.processor.apt;
 
-import static org.jboss.logging.processor.intf.model.Parameter.ParameterType;
+import static org.jboss.logging.processor.model.Parameter.ParameterType;
 import static org.jboss.logging.processor.util.Objects.HashCodeBuilder;
 import static org.jboss.logging.processor.util.Objects.areEqual;
 
@@ -41,9 +41,9 @@ import javax.lang.model.util.ElementFilter;
 import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
 
-import org.jboss.logging.processor.intf.model.MessageMethod;
-import org.jboss.logging.processor.intf.model.Parameter;
-import org.jboss.logging.processor.intf.model.ThrowableType;
+import org.jboss.logging.processor.model.MessageMethod;
+import org.jboss.logging.processor.model.Parameter;
+import org.jboss.logging.processor.model.ThrowableType;
 import org.jboss.logging.processor.util.ElementHelper;
 import org.jboss.logging.processor.util.Objects;
