+++ src/main/java/org/jboss/logging/processor/generator/model/JavaFileObjectCodeWriter.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.logging.processor.model;
+package org.jboss.logging.processor.generator.model;
 
 import java.io.IOException;
 import java.io.OutputStream;
