+++ src/main/java/org/jboss/logging/processor/validation/FormatValidatorFactory.java
@@ -23,7 +23,7 @@
 package org.jboss.logging.processor.validation;
 
 import org.jboss.logging.processor.apt.Annotations.FormatType;
-import org.jboss.logging.processor.intf.model.MessageMethod;
+import org.jboss.logging.processor.model.MessageMethod;
 
 /**
  * Date: 12.08.2011
